Merges some testing facilities based on TL.
