+++ src/FRP/Yampa/Testing.hs
@@ -1,14 +1,23 @@
 module FRP.Yampa.Testing where
 
 import FRP.Yampa.InternalCore
+import FRP.Yampa.Stream
 
--- Testing bit
+-- * Running/initialized SF's
+-- | A wrapper around initialized SF continuations.
 newtype FutureSF a b = FutureSF { unsafeSF :: SF' a b }
 
+-- | Given a signal function and time delta,
+-- it moves the signal function into the future,
+-- returning a new uninitialized SF and the
+-- initial output.
 prefuturize :: SF a b -> a -> DTime -> (SF a b, b)
 prefuturize (SF sf) a dt = let (sf', b) = sf a
   in (SF (sfTF' sf' dt), b)
 
+-- * Evaluation
+
+-- ** One-step evaluation
 evalAtZero :: SF a b
            -> a                  -- Input sample
            -> (b, FutureSF a b)  -- Output x Continuation
@@ -21,9 +30,7 @@ evalAt :: FutureSF a b
 evalAt (FutureSF { unsafeSF = tf }) dt a = (b, FutureSF tf')
   where (tf', b) = (sfTF' tf) dt a
 
-type FutureSampleStream a = [(DTime, a)]
-type SignalSampleStream a = (a, FutureSampleStream a)
-
+-- ** Stream-based evaluation
 evalSF :: SF a b
        -> SignalSampleStream a
        -> (SignalSampleStream b, FutureSF a b)
